// src/lib/types/pocketbase-types.ts
export enum Collections {
	Authorigins = "_authOrigins",
	Externalauths = "_externalAuths",
	Mfas = "_mfas",
	Otps = "_otps",
	Superusers = "_superusers",
	Ingredients = "ingredients",
	Recipes = "recipes",
	RecipesIngredients = "recipes_ingredients",
	Users = "users",
}

// Alias types for improved usability
export type IsoDateString = string
export type IsoAutoDateString = string & { readonly autodate: unique symbol }
export type RecordIdString = string
export type FileNameString = string & { readonly filename: unique symbol }
export type HTMLString = string

type ExpandType<T> = unknown extends T
	? T extends unknown
		? { expand?: unknown }
		: { expand: T }
	: { expand: T }

// System fields
export type BaseSystemFields<T = unknown> = {
	id: RecordIdString
	collectionId: string
	collectionName: Collections
} & ExpandType<T>
// ...
export enum RecipesCategoryOptions {
	"Breakfast" = "Breakfast",
	"Lunch" = "Lunch",
	"Dinner" = "Dinner",
	"Desserts" = "Desserts",
}
export type RecipesRecord<Tinstructions = unknown> = {
	category: RecipesCategoryOptions
	created: IsoAutoDateString
	description?: string
	id: string
	instructions?: null | Tinstructions
	prepTimeMin: number
	recipeCode: string
	title: string
	updated: IsoAutoDateString
}

export type RecipesIngredientsRecord = {
	created: IsoAutoDateString
	id: string
	ingredient: RecordIdString
	quantity: number
	recipe: RecordIdString
	unit?: string
	updated: IsoAutoDateString
}
// ...
export type RecipesResponse<Tinstructions = unknown, Texpand = unknown> = Required<RecipesRecord<Tinstructions>> & BaseSystemFields<Texpand>
export type RecipesIngredientsResponse<Texpand = unknown> = Required<RecipesIngredientsRecord> & BaseSystemFields<Texpand>